Text_Classification
Importing the 20NewsGroups dataset consisting of 11314 articles in the training dataset and 7532 articles in the test dataset accross 20 classes.
Next we Vectorize the articles in the Corpus. For this we use sci-kit learn's CountVectorizer to create a sparse matrix of the count of each word in an article For better results we then calculate the inverse term frequency for the words using sci-kit learn's TfidfTransformer.
Having got the sparse matrix, we would now apply classification Algorithms on this vectorized word matrix to predic classes for data in test dataset. Starting with K-Nearest Neighbours. Next we apply Support Vector Machine algorithm,then we apply Naive Bayes.
All this while we used Bag-Of-Words technique to vectorize the dataset. Here we apply ngrams technique to create the sparse matrix. Let's have an example as how n-grams is differnt from Bag-of-Words and what it actually does.
What if we need to apply Bag-of-2grams, or in other words club two consecutive words in a document, then vectorize.Next we will apply n-grams on our dataset and will then apply SVM classification algorithm.